Making changes about PR and adding messages to all functionality.

+++ academic-record-data/src/main/java/com/java/controller/GradeController.java
@@ -2,71 +2,102 @@ package com.java.controller;
 
 import com.java.dto.GradeDto;
 import com.java.dto.SubjectDto;
+import com.java.model.Grade;
 import com.java.service.GradeService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.Collection;
+import java.util.HashSet;
 
 @RequestMapping("/grade/")
 @RestController
 public class GradeController {
 
+    private static final String ERROR_MESSAGE = "An error occurred: ";
+
     @Autowired
     private GradeService gradeService;
 
     @PostMapping
-    public void save(@Valid @RequestBody GradeDto gradeDto) throws Exception {
+    ResponseEntity<Object> save(@Valid @RequestBody GradeDto gradeDto) {
         try {
             gradeService.save(gradeDto);
+            return new ResponseEntity<>("All good, great created", HttpStatus.CREATED);
         } catch (Exception e){
-            throw new Exception(e);
+            String errorMessage = ERROR_MESSAGE + "Unable to create grade";
+            return new ResponseEntity<>(errorMessage, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @GetMapping("{id}")
-    public Object findById(@PathVariable Integer id) throws Exception {
+    ResponseEntity<Object> findById(@PathVariable Integer id) throws Exception {
         try {
-            return gradeService.findById(id);
+            GradeDto grade = (GradeDto) gradeService.findById(id);
+            if (grade != null){
+                gradeService.findById(id);
+                return new ResponseEntity<>(grade, HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("No grade found", HttpStatus.NOT_FOUND);
+            }
         } catch (Exception e){
-            throw new Exception(e);
+            String errorMessage = ERROR_MESSAGE + "Unable to find grade with id: " + id;
+            return new ResponseEntity<>(errorMessage, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @GetMapping
-    public Collection<GradeDto> findAll() throws Exception {
+    ResponseEntity<Object> findAll() {
         try {
-            return gradeService.findAll();
+            Collection grades = gradeService.findAll();
+            if (grades.isEmpty()) {
+                return new ResponseEntity<>("No grades found", HttpStatus.NOT_FOUND);
+            } else {
+                return new ResponseEntity<>(grades, HttpStatus.OK);
+            }
         } catch (Exception e) {
-            throw new Exception(e);
+            String errorMessage = ERROR_MESSAGE + "Not able to found all grades";
+            return new ResponseEntity<>(errorMessage, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @DeleteMapping("{id}")
-    void delete(@PathVariable Integer id) throws Exception {
+    ResponseEntity<Object> delete(@PathVariable Integer id) {
         try {
-            gradeService.deleteById(id);
+            if (gradeService.findById(id) != null) {
+                gradeService.deleteById(id);
+                return new ResponseEntity<>("Grade with id: " + id + " deleted.", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("Grade does not exist" , HttpStatus.NOT_FOUND);
+            }
         } catch (Exception e){
-            throw new Exception(e);
+            String errorMessage = ERROR_MESSAGE + "Not able to delete grade with id: " + id;
+            return new ResponseEntity<>(errorMessage, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @DeleteMapping
-    void deleteAll() throws Exception {
+    ResponseEntity<Object> deleteAll() {
         try {
             gradeService.deleteAll();
+            return new ResponseEntity<>("All grades were deleted.", HttpStatus.OK);
         } catch (Exception e) {
-            throw new Exception(e);
+            String errorMessage = ERROR_MESSAGE + "Unable to delete all grades";
+            return new ResponseEntity<>(errorMessage, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @PatchMapping("{id}")
-    void update(@Valid @RequestBody GradeDto gradeDto, @PathVariable Integer id) throws Exception {
+    ResponseEntity<Object> update(@Valid @RequestBody GradeDto gradeDto, @PathVariable Integer id) {
         try {
             gradeService.update(gradeDto, id);
+            return new ResponseEntity<>("Grade updated", HttpStatus.OK);
         } catch (Exception e) {
-            throw new Exception(e);
+            String errorMessage = ERROR_MESSAGE + "Not possible to update grade.";
+            return new ResponseEntity<>(errorMessage, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 }
